api/requirements_docker.txt
flask
gdown
numpy
tensorflow_datasets==4.7.0
tensorflow_addons==0.18.0
vit_keras
opencv-python
waitress